Casting fixes for 64 bits. Incomplete commit, discussion on proper casting has to be finished.

+++ source/blender/blenkernel/BKE_customdata.h
@@ -34,7 +34,7 @@
 
 struct CustomData;
 struct CustomDataLayer;
-typedef int CustomDataMask;
+typedef long CustomDataMask;
 
 extern const CustomDataMask CD_MASK_BAREMESH;
 extern const CustomDataMask CD_MASK_MESH;
